fix: get rid of bad uses of ik, pk
`ik` and `pk` should be reserved for the operations like `T[pk]`.  There
are some uses where they serve as variables.  Turn those into plain `i`
and `p`.  Since that is a little more difficult to interpret, later, it
can be clarified that they represent index key and primary key.